+++ bin/check-kube-pods-pending.rb
@@ -30,8 +30,8 @@
 #     --exclude-namespace
 # -i NAMESPACES,                   Include the specified list of namespaces, an
 #     --include-namespace          empty list includes all namespaces
-#     --exclude-node               Exclude the specified list of nodes
-#     --include-node               Include the specified list of nodes, an
+#     --exclude-nodes              Exclude the specified nodes (comma separated list)
+#     --include-nodes              Include the specified nodes (comma separated list), an
 #                                  empty list includes all nodes
 # -t, --timeout TIMEOUT            Threshold for pods to be in the pending state
 # -f, --filter FILTER              Selector filter for pods to be checked
@@ -85,15 +85,15 @@ class AllPodsAreReady < Sensu::Plugins::Kubernetes::CLI
          proc: proc { |a| a.split(',') },
          default: ''
 
-  option :exclude_node,
-         description: 'Exclude the specified list of nodes',
-         long: '--exclude-node NODES',
+  option :exclude_nodes,
+         description: 'Exclude the specified nodes (comma separated list)',
+         long: '--exclude-nodes NODES',
          proc: proc { |a| a.split(',') },
          default: ''
 
-  option :include_node,
-         description: 'Include the specified list of nodes',
-         long: '--include-node NODES',
+  option :include_nodes,
+         description: 'Include the specified nodes (comma separated list)',
+         long: '--include-nodes NODES',
          proc: proc { |a| a.split(',') },
          default: ''
 
@@ -145,7 +145,7 @@ class AllPodsAreReady < Sensu::Plugins::Kubernetes::CLI
   end
 
   def should_exclude_node(node_name)
-    return !config[:include_node].include?(node_name) unless config[:include_node].empty?
-    config[:exclude_node].include?(node_name)
+    return !config[:include_nodes].include?(node_name) unless config[:include_nodes].empty?
+    config[:exclude_nodes].include?(node_name)
   end
 end